% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getterminfo}
\alias{getterminfo}
\title{Retrieve the content of a gene set}
\usage{
getterminfo(i, design, array)
}
\arguments{
\item{i}{The i-th gene set to probe.}

\item{design}{The design matrix.}

\item{array}{The gene-wise information. The order of genes should be the same as in the design matrix.}
}
\value{
A vector containing the result.
}
\description{
Retrieve the content of a gene set
}
