% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hisig_fit}
\alias{hisig_fit}
\title{Fit a linear model with glmnet.}
\usage{
hisig_fit(
  data,
  dummy = 1,
  lambda.min = 0.0001,
  nlambda = 100,
  pos.only = F,
  lambda = NULL,
  random = F,
  simple.output = F
)
}
\arguments{
\item{data}{A named list containing \code{design} and \code{response}.}

\item{dummy}{A dummy variable for parallelzation. Default is 1.}

\item{lambda.min}{A value for \code{lambda.min.ratio} variable of the \code{glmnet}
function.}

\item{nlambda}{A value for \code{nlambda} variable of the \code{glmnet} function.}

\item{pos.only}{If TRUE, force all coefficients to be non-negative. Default
is FALSE.}

\item{lambda}{Provide a vector of lambda when \code{random} is TRUE.}

\item{random}{If TRUE, the fit is for the null model. Deactivate \code{lambda.min}
and \code{nlambda}. Need to provide \code{lambda} value (generated by the main fit).
Default is FALSE.}
}
\value{
When \code{random} is TRUE, only return the \code{beta_max} vector (gene set
impact on responses); when \code{random} is FALSE, return more details of the
fit.
}
\description{
This function takes a "data" object (containing \code{design} and \code{response}), and
fit an entire lasso path. It generates a coefficient matrix of each gene set
at multiple sampled strengths of regularization (\code{lambda}).
}
