% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hisig_fit_rand}
\alias{hisig_fit_rand}
\title{Create the HiSig null model with glmnet.}
\usage{
hisig_fit_rand(
  data,
  lambda,
  batch = 10,
  batch_size = 10,
  n_cores = detectCores() - 1
)
}
\arguments{
\item{data}{A named list containing \code{design} and \code{response}.}

\item{lambda}{An vector of lambda determined by the main fit.}

\item{batch}{Parameter of parallelization. The total number of permutatoin is \code{batch*batch_size}.}

\item{batch_size}{Parameter of parallelization}

\item{n_cores}{Number of cores to use for parallelization}
}
\value{
A matrix to quantify the impact of gene sets under the null model.
}
\description{
Create the HiSig null model with glmnet.
}
