% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{output}
\alias{output}
\title{Save the dataframe.}
\usage{
output(df, out)
}
\arguments{
\item{df}{The dataframe generated by \code{parse_hisig}.}

\item{out}{The output filename.}
}
\value{

}
\description{
Save the dataframe.
}
