% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_hisig}
\alias{parse_hisig}
\title{Parse the HiSig results into a dataframe}
\usage{
parse_hisig(
  data,
  impact,
  term.names,
  gene.names,
  signal2 = NA,
  gene.as.term = F,
  showedge = 20
)
}
\arguments{
\item{data}{A named list containing \code{design} and \code{response}.}

\item{impact}{The impact matrix; 1st column being the main fit; 2nd column to the end being the null model.}

\item{term.names}{a vector of term (gene set) names}

\item{gene.names}{a vector of gene names}

\item{signal2}{if not NA, will print a second numerical information for gene sets. Default is NA.}

\item{gene.as.term}{if the design matrix sees individual gene as a gene set, set it to TRUE to correctly compute statistics. Default is FALSE.}

\item{showedge}{the number of top genes to print when the gene sets are big}
}
\value{
A dataframe, sorted by q values.
}
\description{
Parse the HiSig results into a dataframe
}
