% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_hisig_ms}
\alias{parse_hisig_ms}
\title{Parse HiSig results multi-sample mode}
\usage{
parse_hisig_ms(beta_sample, beta_null)
}
\arguments{
\item{beta_sample}{a matrix (No. regulons)}

\item{beta_null}{The impact matrix; 1st column being the main fit; 2nd column to the end being the null model.}
}
\value{
two dataframes (No. regulons x No. samples). The first one for normalized scores, and the second one for q values. The normalized scores may contain 'NA'.
}
\description{
Parse HiSig results multi-sample mode
}
