% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_hisig_ms}
\alias{parse_hisig_ms}
\title{Parse HiSig results multi-sample mode}
\usage{
parse_hisig_ms(beta_sample, beta_null)
}
\arguments{
\item{beta_sample}{An impact matrix (No. regulons x No. samples)}

\item{beta_null}{An impact matrix (No. r)}
}
\value{
Two dataframes (No. regulons x No. samples). The first one for normalized scores, and the second one for q values. The normalized scores may contain 'NA'.
}
\description{
Parse HiSig results multi-sample mode
}
