% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse2aracne}
\alias{sparse2aracne}
\title{convert the regulon from sparse matrix format to the Aracne/Viper format}
\usage{
sparse2aracne(data)
}
\arguments{
\item{data}{The object created by the 'load_data' function}

\item{design}{The output file name}
}
\value{
regulon A regulon object in viper package
}
\description{
convert the regulon from sparse matrix format to the Aracne/Viper format
}
